refactor(auth): move datasource-user.dto to auth/mappers directory
Update import paths in all affected files and add comprehensive tests for the DTO
