+++ src/infrastructure/external/auth/datasources/__tests__/login.test.ts
@@ -3,7 +3,7 @@ import { AuthClient } from "@/infrastructure/external/auth/auth.client";
 import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { UserEntity } from "@/domain/user/entities/user.entity";
 import { AuthUserEntity } from "@/domain/auth/entities/auth-user.entity";
-import { DatasourceUserDto } from "@/infrastructure/dtos/datasource-user.dto";
+import { DatasourceUserDto } from "@/infrastructure/external/auth/mappers/datasource-user.dto";
 import {
   createMockLoginDto,
   createMockUser,
@@ -17,7 +17,7 @@ import {
 jest.mock("@/infrastructure/external/auth/auth.client");
 jest.mock("@/domain/user/entities/user.entity");
 jest.mock("@/domain/auth/entities/auth-user.entity");
-jest.mock("@/infrastructure/dtos/datasource-user.dto");
+jest.mock("@/infrastructure/external/auth/mappers/datasource-user.dto");
 
 const MockedAuthClient = AuthClient as jest.MockedClass<typeof AuthClient>;
 const MockedUserEntity = UserEntity as jest.MockedClass<typeof UserEntity>;
